feat: image event-bus-pub-sub.png
Develop a visual representation of the Publish-Subscribe model via an event bus. The image should clearly differentiate between the 'publish' and 'subscribe' processes with contrasting colors (such as red for publish, blue for subscribe). Include symbolic representations of data being pushed to a central hub and then distributed to multiple subscribers. Opt for a clean, modern design that effectively communicates the concept of message distribution and asynchronous communication.