Enhancing Fault Recovery in Microservices with Kubernetes (#17)
* feat: new article about efficient fault recovery with Kubernetes in microservices environments

* feat: image kubernetes-fault-recovery-architecture.jpg

Create a detailed diagram showing the architecture of Kubernetes for fault recovery. Illustrate various components like pods, nodes, and their interactions during a fault recovery process. Use clear labels and arrows to show the flow of data and recovery actions. The style should be technical and precise, using a color scheme that differentiates various elements effectively, such as green for active components and red for failed ones.

* feat: image kubernetes-pods-rolling-update.png

Design an image representing a rolling update process in Kubernetes. Show a sequence of pods in different stages of updating, from initiation to completion. Use arrows to indicate the progression and include icons to represent the updating status (like check marks for updated pods and clocks for those in progress). The design should convey a sense of dynamic movement and efficiency, using subtle colors like blues and grays to maintain a professional and technical look.

* feat: image kubernetes-statefulset-diagram.png

Craft a diagram that visualizes the concept of a Kubernetes StatefulSet and its interaction with persistent storage. Highlight elements such as pods, persistent volume claims, and data storage, meticulously labeling each component. Use connecting lines to show relationships and data flow. Opt for a clarifying and educational style with solid colors to distinguish between stateful and stateless components, possibly using shades of orange for stateful elements to emphasize their importance.

* - ✨ Updated the post date to reflect the correct publication timeline.
- 🗑️ Removed an outdated image to improve content relevance.
- ✨ Added a new image illustrating Kubernetes pods rolling update for enhanced understanding.

---------

Co-authored-by: Andy van Dongen <andy.van.dongen@kabisa.nl>